Fixed string cloning.
Simptoms:
FatalExecutionEngineError was detected
Message: The runtime has encountered a fatal error. The address of the error was at 0x6932a0f7, on thread 0xb20. The error code is 0xc0000005. This error may be a bug in the CLR or in the unsafe or non-verifiable portions of user code. Common sources of this bug include user marshaling errors for COM-interop or PInvoke, which may corrupt the stack.